Separate source for mp tests (#5326)
